LangchainService/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
